+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 - Allows for any name to define the subassemblies of a turbine or substation to enable users to use the naming conventions they are most familiar with or most meaningful for their own work
 - Minor bug fixes in the `Metrics` class to improve stability, improve code reuse, and documentation\
 - Adds nearly all documentation udpates from PR #39 as a result of an internal code review, but makes the changes in the source files that generate the example notebooks, so is not a direct merge
+- Adds an annual fee to `PortConfig.annual_fee` that gets applied monthly, though is not included in any metrics as of yet.
+- Adds `UnscheduledServiceEquipmentData.tow_speed` to differentiate between towing speeds and traveling speeds required between port and site, and implements the towing speed application appropriately
+- Adds a `location` flag to the events logging infrastructure and implements its usage across the simulation architecture
+- Creates the metric `Metrics.number_of_tows` to track the number of tows and provides breakdowns as needed
+- Creates the metric `Metrics.vessel_crew_hours_at_sea` to track the number of vessel or crew hours at sea
 
 ## 0.4.1 (2022-March-8)
 - Adds code diagrams to demonstrate how the various components connect